% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_annotation.R
\name{check_f_rm}
\alias{check_f_rm}
\title{Check if any factors for removal are likely native to certain cell types and shouldn't
be removed.}
\usage{
check_f_rm(
  cell.annot,
  f.ct.rm,
  f.from,
  cf.fracs.all,
  ncm.ct.full,
  do.clean = TRUE,
  median.thresh = 0.1
)
}
\arguments{
\item{cell.annot}{character vector of cell types named by cell ids}

\item{f.ct.rm}{list of factors and the corresponding cell types for removal}

\item{f.from}{list of factors and the cell types they are likely sources of admixture from}

\item{cf.fracs.all}{Factor fractions per cell per}

\item{ncm.ct.full}{Nearest neighbor counts from molecules to cell types}

\item{do.clean}{Whether to remove detected false positives or just reformat
the results (default=TRUE)}

\item{median.thresh}{Median expression threshold above which false positives are flagged}
}
\value{
A vector of factor_ctypes suggested for removal and pruned of potential
false positives that are likely to be natively expressed factors.
}
\description{
Check if any factors for removal are likely native to certain cell types and shouldn't
be removed.
}
