% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_annotation.R
\name{check_f_rm_per_ct}
\alias{check_f_rm_per_ct}
\title{Check for false positive removals for NMFs done per cell type}
\usage{
check_f_rm_per_ct(
  cell.annot,
  mb.res,
  ct.fracs.per.ct,
  ncm.ct.full,
  do.clean = TRUE,
  median.thresh = 0.1
)
}
\arguments{
\item{cell.annot}{character vector of cell types named by cell ids}

\item{mb.res}{List of suggested removals per cell type}

\item{ct.fracs.per.ct}{Factor fractions per cell per cell type}

\item{ncm.ct.full}{Nearest neighbor counts from molecules to cell types}

\item{do.clean}{Whether to remove detected false positives or just reformat
the results (default=TRUE)}

\item{median.thresh}{Median expression threshold above which false positives are flagged}
}
\value{
A vector of factor_ctypes suggested for removal and pruned of potential
false positives that are likely to be natively expressed factors.
}
\description{
Check for false positive removals for NMFs done per cell type
}
