% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_scores.R
\name{estimate_cell_admixture_fractions}
\alias{estimate_cell_admixture_fractions}
\title{Estimate cell contamination fractions}
\usage{
estimate_cell_admixture_fractions(gene.cont.probs, cell.type.cm, mask = NULL)
}
\description{
Estimates \eqn{E(f_c | ...)}
}
