% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_scores.R
\name{estimate_contamination_scores}
\alias{estimate_contamination_scores}
\title{Title}
\usage{
estimate_contamination_scores(
  cm.rna,
  cm.spatial,
  annot.rna,
  annot.spatial,
  cell.type.adj.mat,
  p.c = 0.25,
  signal.thres = 0.25,
  min.expr.frac = 0.05,
  max.prob.high = 0.1,
  exclude.cell.types = NULL,
  exclude.genes = NULL,
  use.counts = FALSE,
  adjust = TRUE
)
}
\arguments{
\item{cm.rna}{counts matrix for scRNA-seq data}

\item{cm.spatial}{counts matrix for spatial data}

\item{annot.rna}{cell-type annotation for sc}

\item{annot.spatial}{cell-type annotation for spatial}

\item{cell.type.adj.mat}{ctype adjacency matrix}

\item{p.c}{prior probability of contamination}

\item{signal.thres}{threshold for selecting admixture genes}

\item{min.expr.frac}{min expression fraction}

\item{max.prob.high}{max prob}

\item{exclude.cell.types}{cell types to exclude}

\item{exclude.genes}{genes to exclude}

\item{use.counts}{whether to use counts (default=FALSE)}

\item{adjust}{whether to adjust (default=TRUE)}
}
\value{
contamination scores
}
\description{
Title
}
