% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_scores.R
\name{estimate_contamination_scores_seurat}
\alias{estimate_contamination_scores_seurat}
\title{Estimate contamination scores from Seurat objects}
\usage{
estimate_contamination_scores_seurat(
  so.rna,
  so.spatial,
  cell.type.adj.mat,
  idents = "cell_type",
  ...
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=estimate_contamination_scores]{estimate_contamination_scores}}
  \describe{
    \item{\code{cm.rna}}{counts matrix for scRNA-seq data}
    \item{\code{cm.spatial}}{counts matrix for spatial data}
    \item{\code{annot.rna}}{cell-type annotation for sc}
    \item{\code{annot.spatial}}{cell-type annotation for spatial}
    \item{\code{cell.type.adj.mat}}{ctype adjacency matrix}
    \item{\code{p.c}}{prior probability of contamination}
    \item{\code{signal.thres}}{threshold for selecting admixture genes}
    \item{\code{min.expr.frac}}{min expression fraction}
    \item{\code{max.prob.high}}{max prob}
    \item{\code{exclude.cell.types}}{cell types to exclude}
    \item{\code{exclude.genes}}{genes to exclude}
    \item{\code{use.counts}}{whether to use counts (default=FALSE)}
    \item{\code{adjust}}{whether to adjust (default=TRUE)}
  }}
}
\description{
Estimate contamination scores from Seurat objects
}
