% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_scores.R
\name{estimate_gene_contamination_probabilities}
\alias{estimate_gene_contamination_probabilities}
\title{Estimate gene contamination probabilities}
\usage{
estimate_gene_contamination_probabilities(
  gene.scores,
  p.c = NULL,
  prior.cont.prob = NULL
)
}
\arguments{
\item{gene.scores}{Data.frame with columns \code{cont} = \eqn{\sum_{v \neq u} \left(p(g_i | t_v, M) \cdot \frac{k_{v,u}}{\sum_{x \neq u} k_{x, u}} \right)} and \code{ref}=\eqn{p(g_i | t_u, M)}}

\item{p.c}{Proportion of cells in the contamination cluster \eqn{p_c}. If null, forces \eqn{p_{c,u} = 0.5}}

\item{prior.cont.prob}{Prior probability of contamination \eqn{\frac{\sum_{x \neq u}k_{x,u}}{\sum_{x}k_{x,u}}}}
}
\description{
Estimates \eqn{p(m_i \in C | g_i, ...)}
}
