% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admix_scores.R
\name{estimate_gene_scores}
\alias{estimate_gene_scores}
\title{Estimate gene scores}
\usage{
estimate_gene_scores(prob.gene.per.type, cell.type.adj.mat, cell.type)
}
\arguments{
\item{prob.gene.per.type}{Gene probabilities per cell type \eqn{p(g_i | t_x, M)}}

\item{cell.type.adj.mat}{Cell type adjacency matrix \eqn{K}}

\item{cell.type}{Current cell type \eqn{u}}
}
\value{
Data.frame with columns \code{cont} = \eqn{\sum_{v \neq u} \left(p(g_i | t_v, M) \cdot \frac{k_{v,u}}{\sum_{x \neq u} k_{x, u}} \right)} and \code{ref}=\eqn{p(g_i | t_u, M)}
}
\description{
Estimates terms of the Bayesian formula for  \eqn{p(m_i \in C | g_i, ...)}
}
