% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_annotation.R
\name{extract_bridge_res}
\alias{extract_bridge_res}
\title{Extract results from bridge test result into a list of factor-ctype pairs for removal}
\usage{
extract_bridge_res(
  combo.res,
  all.ctypes,
  n.factors,
  nmf.type = "joint",
  p.thresh = 0.01,
  adj.pvals = TRUE,
  get.p.mat = FALSE
)
}
\arguments{
\item{combo.res}{List of results from bridge tests}

\item{all.ctypes}{Vector of cell types}

\item{n.factors}{Number of factors tested}

\item{nmf.type}{Character whether the cell type was run jointly with all cell types
together or separately per cell type ('joint' or 'ct')}

\item{p.thresh}{pvalue threshold below which to select factor-ct combinations for removal}

\item{adj.pvals}{Whether to adjust pvalues with FDR or not}

\item{get.p.mat}{Whether to get the p-value matrix instead of the list of pairs}
}
\value{
either a list of factor-ctype pairs or a matrix of pvalues.
}
\description{
Extract results from bridge test result into a list of factor-ctype pairs for removal
}
