% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_counts_meta_seurat}
\alias{get_counts_meta_seurat}
\title{Transform transcript molecule df into standard seurat object with counts matrix}
\usage{
get_counts_meta_seurat(tx.dat, normalize = FALSE)
}
\arguments{
\item{tx.dat}{dataframe of transcripts with at least gene, cell, and celltype columns}
}
\value{
Seurat object with the data in counts matrix form
}
\description{
Transform transcript molecule df into standard seurat object with counts matrix
}
