% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_annotation.R
\name{get_enr}
\alias{get_enr}
\title{Run the enrichment test for each cell type's markers in each factor}
\usage{
get_enr(
  res,
  markers,
  n.perm = 10000,
  p.thresh = 0.1,
  adj.pvals = TRUE,
  crf.h = NULL,
  df = NULL,
  n.cores = 1
)
}
\arguments{
\item{res}{NMF results object}

\item{markers}{Data Frame of markers with columns 'gene' and 'marker_of'}

\item{n.perm}{Number of permutations to run (default=10000)}

\item{p.thresh}{Threshold below which to select significant results for removal}

\item{adj.pvals}{Whether to adjust pvalues using FDR (default=TRUE)}

\item{crf.h}{The Neighborhood size used in CRF (default=NULL)}

\item{df}{Data Frame with molecule cordinates, mol_id, gene, cell, celltype columns (default=NULL)}

\item{n.cores}{Number of cores to run parallel with (default=1)}
}
\value{
A list of factor cell type combinations for removal
}
\description{
Run the enrichment test for each cell type's markers in each factor
}
