% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam.R
\name{get_graph}
\alias{get_graph}
\title{Get graph from cell}
\usage{
get_graph(df.cell, h = 10, distinct.edges = FALSE)
}
\arguments{
\item{h}{Neighborhood size}

\item{df_cell}{Data frame with molecule coordinates}

\item{distinct_edges}{Whether to keep only distinct edges}
}
\value{
Graph object
}
\description{
Get graph from cell
}
