% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam.R
\name{get_knn_counts}
\alias{get_knn_counts}
\title{Get KNN counts for each molecule}
\usage{
get_knn_counts(df, k, include_i = FALSE, ncores = 1)
}
\arguments{
\item{df}{Data frame with molecule coordinates}

\item{k}{Neighborhood size}

\item{ncores}{Number of cores to use}
}
\value{
Data frame with KNN counts for each molecule
}
\description{
Get KNN counts for each molecule
}
