% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam.R
\name{get_knn_counts_all}
\alias{get_knn_counts_all}
\title{Get KNN count matrices for molecules in all cells separately}
\usage{
get_knn_counts_all(df, h, include_i = FALSE, n.cores = 1)
}
\arguments{
\item{df}{Data frame with molecule coordinates}

\item{h}{Neighborhood size}

\item{n.cores}{Number of cores to use}
}
\value{
Data frame with KNN counts for each molecule
}
\description{
Get KNN count matrices for molecules in all cells separately
}
