% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam.R
\name{knn_adjacency_matrix}
\alias{knn_adjacency_matrix}
\title{Create and adjacency matrix}
\usage{
knn_adjacency_matrix(df, k, weighted = FALSE)
}
\arguments{
\item{df}{Data frame with molecule coordinates}

\item{k}{Neighborhood size}

\item{weighted}{Logical whether to create a weighted matrix}
}
\value{
sparse matrix of molecule adjacencies
}
\description{
Create and adjacency matrix
}
