% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam.R
\name{knn_count_matrix}
\alias{knn_count_matrix}
\title{Get KNN counts for each molecule}
\usage{
knn_count_matrix(df, k, all.genes = NULL, include_i = FALSE)
}
\arguments{
\item{df}{Data frame with molecule coordinates}

\item{k}{Neighborhood size}

\item{all.genes}{all genes to include as columns (defualt is all genes present in df)}

\item{include_i}{whether to include self-molecule gene as a nearest neighbor}
}
\value{
a molecule by gene matrix of knn counts
}
\description{
Get KNN counts for each molecule
}
