% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_annotation.R
\name{normalize_images}
\alias{normalize_images}
\title{Normalize images for membrane score tests}
\usage{
normalize_images(im)
}
\arguments{
\item{im}{image matrix with dimensions of z by y by x}
}
\value{
the normalized image
}
\description{
Normalize images for membrane score tests
}
