% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_annot_hmap}
\alias{plot_annot_hmap}
\title{Plot p-value results from bridge- or membrane-based factor annotation}
\usage{
plot_annot_hmap(
  rs,
  ad.sources,
  final.rm,
  annot.type,
  p.bridge = 0.1,
  p.memb = 0.01
)
}
\arguments{
\item{rs}{The p-value rowmax matrix for admixture into each cell types.
Dims are cell types x factors}

\item{ad.sources}{A list of reported "admixture-from" cell types for each factor}

\item{final.rm}{The vector of factor cell type combinations for removal, structured
as f_ctype.}

\item{annot.type}{Either 'membrane' or 'bridge' depending on the type of annotation
used to generate the other inputs.}

\item{p.bridge}{The p-value used to extract bridge results}

\item{p.memb}{The p-value used to extract membrane results}
}
\value{
a ComplexHeatmap object showing p-values, as well as which cell types are
recommended for removal (labeled *) and which source cell types are behind each
factor (labeled S)
}
\description{
Plot p-value results from bridge- or membrane-based factor annotation
}
