% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cell_score_ratios}
\alias{plot_cell_score_ratios}
\title{Plot admixture score ratios (P_admix clean)/(P_admix unclean) per cell.
Only implemented for joint NMF/CRF results currently.}
\usage{
plot_cell_score_ratios(
  scores.orig,
  scores.cln,
  f.rm,
  df,
  crf.res,
  upper.lim = 1.5,
  min.mean.thresh = 0.001,
  add.theme = theme_get()
)
}
\arguments{
\item{scores.orig}{Admixture scores data.frame for uncleaned cells}

\item{scores.cln}{Admixture scores data.frame for cleaned cells}

\item{f.rm}{Character vector of factor-ctype combinations to remove}

\item{df}{Data Frame of molecule coordinates, gene, cell, and mol_id columns}

\item{crf.res}{Data Frame of CRF molecule assignments (default=.9)}

\item{upper.lim}{Upper limit of score ratios to show (default=1.5)}

\item{min.mean.thresh}{Lower threshold, below which cell types are excluded}

\item{add.theme}{ggplot2 theme to add to the plots (default: current theme)}
}
\value{
a list of 4 plots including: admixture scores, score ratios, \% cleaned,
and number of molecules before/after cleaning
}
\description{
Plot admixture score ratios (P_admix clean)/(P_admix unclean) per cell.
Only implemented for joint NMF/CRF results currently.
}
