% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_correlation_preservation}
\alias{plot_correlation_preservation}
\title{Plot the preservation of gene correlation}
\usage{
plot_correlation_preservation(cor.cors, bins = 30)
}
\arguments{
\item{cor.cors}{A vector of correlation values}

\item{bins}{The number of bins to use in the histogram}
}
\value{
A ggplot object
}
\description{
Plot the preservation of gene correlation
}
