% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_nmf_loadings}
\alias{plot_nmf_loadings}
\title{Plot top NMF loadings for each factor, optionally also showing clean scRNA expression}
\usage{
plot_nmf_loadings(
  nmf.res,
  ct.exp = NULL,
  f.show = NULL,
  nplt.row = 2,
  add.theme = theme_get()
)
}
\arguments{
\item{nmf.res}{NMF results object}

\item{ct.exp}{Matrix of scaled scRNA expression fractions (default=NULL)}

\item{f.show}{Vector of factors to plot. If NULL, will plot all (default=NULL)}

\item{nplt.row}{Number of rows to place all the heatmaps on (default=2)}
}
\value{
a list of ComplexHeatmap objects
}
\description{
Plot top NMF loadings for each factor, optionally also showing clean scRNA expression
}
