% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_annotation.R
\name{run_bridge_test}
\alias{run_bridge_test}
\title{Run cell-bridging test to determine factors that represent admixture into various cell types}
\usage{
run_bridge_test(
  df,
  crf.f.labs,
  cell.annot,
  ncm.ct,
  ncm.cell,
  ncells.samp = 200,
  knn.k = 20,
  n.cores = 1
)
}
\arguments{
\item{df}{Data Frame of transcript locations, mol_id, cell, and celltype columns}

\item{crf.f.labs}{Data Frame of CRF results}

\item{cell.annot}{Data Frame of spatial metadata}

\item{ncm.ct}{Matrix of all KNN counts for molecules to cell types}

\item{ncm.cell}{Matrix of all KNN counts for molecules to cells}

\item{ncells.samp}{Number of cells to subsample to run the test with for speed (default=200)}

\item{knn.k}{Number of nearest neighbors to use for KNN (default=20)}

\item{n.cores}{Number of cores to run parallel with (default=1)}
}
\value{
a list of p-values and top quantiles for each pair of cell types tested
}
\description{
Run cell-bridging test to determine factors that represent admixture into various cell types
}
