% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam.R
\name{run_crf}
\alias{run_crf}
\title{Run CRF for a cell}
\usage{
run_crf(df.cell, lds, k, h = 10, t = 0.05, max.iter = 10000)
}
\arguments{
\item{lds}{Pre-normalized NMF loadings}

\item{k}{Number of factors}

\item{h}{Neighborhood size}

\item{t}{Transition probability}

\item{df_cell}{Data frame with coordinates and gene labels for one cell}
}
\value{
Data frame with MRF segmentation
}
\description{
Run CRF for a cell
}
