% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam.R
\name{run_crf_all}
\alias{run_crf_all}
\title{Run crf for all cells}
\usage{
run_crf_all(
  df,
  res,
  num.nn = 10,
  same.label.ratio = 5,
  normalize.by = c("gene", "factor", "gene.factor"),
  proj.h = NULL,
  n.cores = 1
)
}
\arguments{
\item{df}{Data Frame with molecule coordinates, gene, cell, and mol_id columns}

\item{res}{NMF results object}

\item{same.label.ratio}{Transition ratio P_no_change/P_to_FactorX (default=5)}

\item{normalize.by}{Method to use for normalizing NMF loadings (default='gene')}

\item{n.cores}{Number of cores to run parallel with (default=1)}

\item{num_nn}{Neighborhood size to use in knn for CRF (default=10)}
}
\value{
A Data Frame with a column for molecule factor assignment, ordered same as df
}
\description{
Run crf for all cells
}
