% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam.R
\name{run_knn_nmf}
\alias{run_knn_nmf}
\title{Run knn for each cell and NMF using molecules from many cells}
\usage{
run_knn_nmf(
  df,
  k = 5,
  h = 20,
  mols.test = NULL,
  nmol.dsamp = NULL,
  n.cores = 1
)
}
\arguments{
\item{df}{Data frame with molecule coordinates}

\item{k}{Number of factors (default=5)}

\item{h}{Neighborhood size for KNN (default=20)}

\item{n.cores}{Number of cores to run parallel with (defualt=1)}

\item{mols_test}{Specific molecules to test (default=NULL)}

\item{nmol_dsamp}{Number of molecules to downsample to (default=NULL)}
}
\value{
the NMF results object
}
\description{
Run knn for each cell and NMF using molecules from many cells
}
