% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_annotation.R
\name{run_memb_test}
\alias{run_memb_test}
\title{Run membrane test to determine factor-ctype pairs for removal}
\usage{
run_memb_test(
  df,
  crf.f.labs,
  im.load.fn,
  num.factors,
  min.mols = 5,
  max.cells.per.fov = 10,
  knn.k = 10,
  all.fov = NULL,
  use.block.knn = FALSE,
  c.dist.expand = 0.025,
  ncm.ct = NULL,
  ncm.cell = NULL,
  n.cores = 1
)
}
\arguments{
\item{df}{Data Frame of molecule coordinates, with mol_id, cell, gene columns}

\item{crf.f.labs}{Data Frame of CRF results}

\item{im.load.fn}{A function to load the images for a give field of view or the whole dataset}

\item{num.factors}{Number of factors}

\item{min.mols}{Minimum number of admixture molecules per cell (default=5)}

\item{max.cells.per.fov}{Maximum number of cells to test per field of view (default=10)}

\item{knn.k}{Neighborhood size for knn graph used to find adjacent cells (default=10)}

\item{all.fov}{Vector of field of views to test or NULL if all should be tested (default=NULL)}

\item{use.block.knn}{Logical whether to run the KNN in blocks over the image.
This can be helpful to reduce runtime for large datasets. (default=FALSE)}

\item{c.dist.expand}{A fraction of the biggest molecule-centroid distance. Used to
select random molecules against for comparison based on current molecule
centroid dist +- the expand. May need to increase this for datasets with less
coverage (default=.025).}

\item{ncm.ct}{knn counts matrix from molecules to cell types. Can only input if
there is a single fov for the whole dataset (default=NULL)}

\item{ncm.cell}{knn counts matrix from molecules to cells. Can only input if
there is a single fov for the whole dataset (default=NULL)}

\item{n.cores}{Number of cores to run parallel with (default=1)}
}
\value{
A list with the membrane test scores per cell for each ctype comparison
}
\description{
Run membrane test to determine factor-ctype pairs for removal
}
