% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam.R
\name{samp_ct_equal}
\alias{samp_ct_equal}
\title{Downsample cells to balance their representation in the dataset}
\usage{
samp_ct_equal(df, cell.annot, num.cells.samp)
}
\arguments{
\item{df}{Data Frame of molecule coordinates, gene, cell, and celltype columns}

\item{cell_annot}{Data Frame of metadata}

\item{num_cells_samp}{Approximately the total number of cells to sample in the dataset}
}
\value{
the df downsampled to the selected cells
}
\description{
Downsample cells to balance their representation in the dataset
}
