% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam.R
\name{subset_genes}
\alias{subset_genes}
\title{Downsample the data to a subset of genes before doing NMF}
\usage{
subset_genes(df, cell.annot, top.g.thresh = NULL, n.hvgs = NULL)
}
\arguments{
\item{df}{Data Frame of molecule coordinates, gene, cell, and celltype columns}

\item{cell_annot}{Data Frame of metadata}

\item{top_g_thresh}{Numeric between 0-1. Only select genes with mean fractional
expression above this threshold in any cell type (default=NULL)}

\item{n_hvgs}{Numeric number of high variance genes to select (default=NULL)}
}
\value{
the df subset to the selected genes.
}
\description{
Downsample the data to a subset of genes before doing NMF
}
