% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam.R
\name{knn_count_matrix}
\alias{knn_count_matrix}
\title{This is a much faster implementation than \code{knn_count_mat}}
\usage{
knn_count_matrix(df, k, all.genes = NULL)
}
\description{
This is a much faster implementation than \code{knn_count_mat}
}
