% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{prep_sc_for_plt}
\alias{prep_sc_for_plt}
\title{Get scaled average expression fractions per cell type from scRNA data}
\usage{
prep_sc_for_plt(sc.obj, cell.type.col = NULL)
}
\arguments{
\item{sc.obj}{sc.obj is seurat object with a cell_type meta.data column}
}
\value{
matrix of z-transformed mean expression fractions per cell type
}
\description{
Get scaled average expression fractions per cell type from scRNA data
}
