% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontam.R
\name{run_crf}
\alias{run_crf}
\title{Run MRF}
\usage{
run_crf(df_cell, expr, h = 10, t = 0.05, verbose = TRUE, ...)
}
\arguments{
\item{df_cell}{Data frame with molecule coordinates}

\item{expr}{Data frame with gene expression}

\item{h}{Neighborhood size}

\item{t}{Transition probability}

\item{k}{Number of factors}
}
\value{
Data frame with MRF segmentation
}
\description{
Run MRF
}
